Adding more Linear Algebra Tests
Additionally, fixed a small bug in the .inv() method. While .inv()
returned the correct inverse, it did so with an additional column at
index '-1'. Although the other functions in the library ignored the '-1'
column, the additinal column prevented equality comparisons. This has
been fixed.

The tests for .inv() are incomplete, pending a deep-approximate equality
method. (See mailing list)